[IMP] fs_attachement, fs_storage: Move storage to model / field association to base addon
