+++ fs_attachment/models/__init__.py
@@ -2,5 +2,3 @@ from . import fs_file_gc
 from . import fs_storage
 from . import ir_attachment
 from . import ir_binary
-from . import ir_model
-from . import ir_model_fields
