+++ fs_attachment/tests/test_fs_storage.py
@@ -3,8 +3,6 @@
 import base64
 import os
 
-from odoo.exceptions import ValidationError
-
 from .common import TestFSAttachmentCommon
 
 
@@ -14,87 +12,6 @@ class TestFsStorage(TestFSAttachmentCommon):
         res = super().setUpClass()
         cls.default_backend = cls.env.ref("fs_storage.fs_storage_demo")
         return res
-
-    def test_compute_model_ids(self):
-        """
-        Give a list of model xmlids and check that the o2m field model_ids
-        is correctly fulfilled.
-        """
-        self.temp_backend.model_xmlids = (
-            "base.model_res_partner,base.model_ir_attachment"
-        )
-
-        model_ids = self.temp_backend.model_ids
-        self.assertEqual(len(model_ids), 2)
-        model_names = model_ids.mapped("model")
-        self.assertEqual(set(model_names), {"res.partner", "ir.attachment"})
-
-    def test_inverse_model_ids(self):
-        """
-        Modify backend model_ids and check the char field model_xmlids
-        is correctly updated
-        """
-        model_1 = self.env["ir.model"].search([("model", "=", "res.partner")])
-        model_2 = self.env["ir.model"].search([("model", "=", "ir.attachment")])
-        self.temp_backend.model_ids = [(6, 0, [model_1.id, model_2.id])]
-        self.assertEqual(
-            self.temp_backend.model_xmlids,
-            "base.model_res_partner,base.model_ir_attachment",
-        )
-
-    def test_compute_field_ids(self):
-        """
-        Give a list of field xmlids and check that the o2m field field_ids
-        is correctly fulfilled.
-        """
-        self.temp_backend.field_xmlids = (
-            "base.field_res_partner__id,base.field_res_partner__create_date"
-        )
-
-        field_ids = self.temp_backend.field_ids
-        self.assertEqual(len(field_ids), 2)
-        field_names = field_ids.mapped("name")
-        self.assertEqual(set(field_names), {"id", "create_date"})
-        field_models = field_ids.mapped("model")
-        self.assertEqual(set(field_models), {"res.partner"})
-
-    def test_inverse_field_ids(self):
-        """
-        Modify backend field_ids and check the char field field_xmlids
-        is correctly updated
-        """
-        field_1 = self.env["ir.model.fields"].search(
-            [("model", "=", "res.partner"), ("name", "=", "id")]
-        )
-        field_2 = self.env["ir.model.fields"].search(
-            [("model", "=", "res.partner"), ("name", "=", "create_date")]
-        )
-        self.temp_backend.field_ids = [(6, 0, [field_1.id, field_2.id])]
-        self.assertEqual(
-            self.temp_backend.field_xmlids,
-            "base.field_res_partner__id,base.field_res_partner__create_date",
-        )
-
-    def test_constraint_unique_storage_model(self):
-        """
-        A given model can be linked to a unique storage
-        """
-        self.temp_backend.model_xmlids = (
-            "base.model_res_partner,base.model_ir_attachment"
-        )
-        self.env.ref("fs_storage.fs_storage_demo")
-        with self.assertRaises(ValidationError):
-            self.default_backend.model_xmlids = "base.model_res_partner"
-
-    def test_constraint_unique_storage_field(self):
-        """
-        A given field can be linked to a unique storage
-        """
-        self.temp_backend.field_xmlids = (
-            "base.field_res_partner__id,base.field_res_partner__name"
-        )
-        with self.assertRaises(ValidationError):
-            self.default_backend.field_xmlids = "base.field_res_partner__name"
 
     def test_force_model_create_attachment(self):
         """
