+++ fs_attachment/views/fs_storage.xml
@@ -8,18 +8,6 @@
         <field name="inherit_id" ref="fs_storage.fs_storage_form_view" />
         <field name="arch" type="xml">
             <field name="options" position="after">
-                <field name="model_ids" widget="many2many">
-                    <list>
-                        <field name="name" />
-                        <field name="model" />
-                    </list>
-                </field>
-                <field name="field_ids" widget="many2many">
-                    <list>
-                        <field name="name" />
-                        <field name="model" />
-                    </list>
-                </field>
                 <separator string="Attachment" />
                 <field name="optimizes_directory_path" />
                 <field name="autovacuum_gc" />
