+++ fs_storage/README.rst
@@ -292,6 +292,7 @@ Contributors
 
 - Laurent Mignon <laurent.mignon@acsone.eu>
 - Sébastien BEAU <sebastien.beau@akretion.com>
+- Marie Lejeune <marie.lejeune@acsone.eu>
 
 Maintainers
 -----------
