+++ fs_storage/models/__init__.py
@@ -1,1 +1,3 @@
 from . import fs_storage
+from . import ir_model
+from . import ir_model_fields
