+++ fs_storage/tests/test_fs_storage.py
@@ -3,6 +3,7 @@
 import warnings
 from unittest import mock
 
+from odoo.exceptions import ValidationError
 from odoo.tests import Form
 from odoo.tools import mute_logger
 
@@ -10,6 +11,18 @@ from .common import TestFSStorageCase
 
 
 class TestFSStorage(TestFSStorageCase):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.copy_backend = cls.env["fs.storage"].create(
+            {
+                "name": "Temp FS Storage",
+                "protocol": "file",
+                "code": "copy_tmp_dir",
+                "directory_path": "/",
+            }
+        )
+
     @mute_logger("py.warnings")
     def _test_deprecated_setting_and_getting_data(self):
         # Check that the directory is empty
@@ -150,3 +163,80 @@ class TestFSStorage(TestFSStorageCase):
             ),
             log.output[0],
         )
+
+    def test_compute_model_ids(self):
+        """
+        Give a list of model xmlids and check that the o2m field model_ids
+        is correctly fulfilled.
+        """
+        self.backend.model_xmlids = "base.model_res_partner,base.model_ir_attachment"
+
+        model_ids = self.backend.model_ids
+        self.assertEqual(len(model_ids), 2)
+        model_names = model_ids.mapped("model")
+        self.assertEqual(set(model_names), {"res.partner", "ir.attachment"})
+
+    def test_inverse_model_ids(self):
+        """
+        Modify backend model_ids and check the char field model_xmlids
+        is correctly updated
+        """
+        model_1 = self.env["ir.model"].search([("model", "=", "res.partner")])
+        model_2 = self.env["ir.model"].search([("model", "=", "ir.attachment")])
+        self.backend.model_ids = [(6, 0, [model_1.id, model_2.id])]
+        self.assertEqual(
+            self.backend.model_xmlids,
+            "base.model_res_partner,base.model_ir_attachment",
+        )
+
+    def test_compute_field_ids(self):
+        """
+        Give a list of field xmlids and check that the o2m field field_ids
+        is correctly fulfilled.
+        """
+        self.backend.field_xmlids = (
+            "base.field_res_partner__id,base.field_res_partner__create_date"
+        )
+
+        field_ids = self.backend.field_ids
+        self.assertEqual(len(field_ids), 2)
+        field_names = field_ids.mapped("name")
+        self.assertEqual(set(field_names), {"id", "create_date"})
+        field_models = field_ids.mapped("model")
+        self.assertEqual(set(field_models), {"res.partner"})
+
+    def test_inverse_field_ids(self):
+        """
+        Modify backend field_ids and check the char field field_xmlids
+        is correctly updated
+        """
+        field_1 = self.env["ir.model.fields"].search(
+            [("model", "=", "res.partner"), ("name", "=", "id")]
+        )
+        field_2 = self.env["ir.model.fields"].search(
+            [("model", "=", "res.partner"), ("name", "=", "create_date")]
+        )
+        self.backend.field_ids = [(6, 0, [field_1.id, field_2.id])]
+        self.assertEqual(
+            self.backend.field_xmlids,
+            "base.field_res_partner__id,base.field_res_partner__create_date",
+        )
+
+    def test_constraint_unique_storage_model(self):
+        """
+        A given model can be linked to a unique storage
+        """
+        self.backend.model_xmlids = "base.model_res_partner,base.model_ir_attachment"
+        self.env.ref("fs_storage.fs_storage_demo")
+        with self.assertRaises(ValidationError):
+            self.copy_backend.model_xmlids = "base.model_res_partner"
+
+    def test_constraint_unique_storage_field(self):
+        """
+        A given field can be linked to a unique storage
+        """
+        self.backend.field_xmlids = (
+            "base.field_res_partner__id,base.field_res_partner__name"
+        )
+        with self.assertRaises(ValidationError):
+            self.copy_backend.field_xmlids = "base.field_res_partner__name"
