+++ fs_storage/views/fs_storage_view.xml
@@ -42,6 +42,18 @@
                                 options="{'mode': 'python'}"
                                 placeholder="Enter you fsspec options here."
                             />
+                            <field name="model_ids" widget="many2many">
+                                <list>
+                                    <field name="name" />
+                                    <field name="model" />
+                                </list>
+                            </field>
+                            <field name="field_ids" widget="many2many">
+                                <list>
+                                    <field name="name" />
+                                    <field name="model" />
+                                </list>
+                            </field>
                             <field name="check_connection_method" />
                         </group>
                         <group>
